fix(mcp-server): SMI-5991 stub getInstalledSkills in recommend.test.ts
recommend.test.ts (the older, standalone file, distinct from
recommend-online-path.test.ts) never set project_context and relied on
real auto-detection against the host's actual ~/.claude/skills/
directory. Most of the file's installed_skills: [] cases depend on that
auto-detection returning something non-empty to populate the derived
context.stack; when it returns empty (e.g. any clean Docker container or
CI runner — confirmed empirically: /root/.claude/skills/ doesn't exist
in a fresh worktree container), the SMI-5896 empty-stack guard fires and
short-circuits to a degraded response with recommendations: [] —
silently turning most assertions in this file into vacuous passes and
failing the two that actually check for non-empty recommendations.

Stub getInstalledSkills to a fixed, synthetic, non-colliding skill ID
(not a real TEST_SKILLS fixture like 'anthropic/commit' — verified that
choice first, since with detect_overlap defaulted true it engages
OverlapDetector against real fixture data and can legitimately zero
every candidate). This makes the derived stack deterministic across
every host/container/CI environment without perturbing the
installed-skill exclusion or overlap-detection paths other tests in
this file exercise deliberately via an explicit installed_skills array.

The fix pushed the file to 524 lines (over the 500-line cap), so
formatRecommendations tests are split out to recommend-format.test.ts —
same precedent as recommend-online-path.test.ts (SMI-2755 Wave 2).

Verified: 44/44 tests pass across all three recommend-* test files, 4
consecutive runs; full mcp-server suite (2853 tests) shows no other
regressions.

Fixes SMI-5991
